+++ src/features/auth/signup/components/SignupPageClient.test.tsx
@@ -2,7 +2,7 @@ import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-import SignupPageClient from "@/app/(auth)/signup/SignupPageClient";
+import SignupPageClient from "@/features/auth/signup/components/SignupPageClient";
 
 const mockPush = vi.fn();
 const mockMutateAsync = vi.fn();
